+++ chainstate/tx-verifier/src/transaction_verifier/pos_accounting_undo_cache/mod.rs
@@ -214,3 +214,6 @@ impl PoSAccountingBlockUndoCache {
         Ok(())
     }
 }
+
+#[cfg(test)]
+mod tests;
