+++ chainstate/tx-verifier/src/transaction_verifier/pos_accounting_undo_cache/tests.rs
@@ -15,12 +15,21 @@
 
 use super::*;
 
-use common::{chain::config::Builder as ConfigBuilder, primitives::H256};
-use mockall::predicate::eq;
-use pos_accounting::DeltaMergeUndo;
-use rstest::rstest;
+use crate::{
+    transaction_verifier::{flush, storage::TransactionVerifierStorageRef, tests::mock},
+    TransactionVerifier,
+};
+
+use common::{
+    chain::{config::Builder as ConfigBuilder, Block},
+    primitives::H256,
+};
+use pos_accounting::{DeltaMergeUndo, TxUndo};
 use test_utils::random::Seed;
-use tokens_accounting::{TokensAccountingDeltaUndoData, TxUndo};
+use tokens_accounting::TokensAccountingDeltaUndoData;
+
+use mockall::predicate::eq;
+use rstest::rstest;
 
 #[rstest]
 #[trace]
@@ -30,10 +39,10 @@ fn connect_txs_in_hierarchy_default(#[case] seed: Seed) {
 
     let chain_config = ConfigBuilder::test_chain().build();
 
-    let undo1 = tokens_accounting::random_undo_for_test(&mut rng);
-    let undo2 = tokens_accounting::random_undo_for_test(&mut rng);
-    let undo3 = tokens_accounting::random_undo_for_test(&mut rng);
-    let undo3_2 = tokens_accounting::random_undo_for_test(&mut rng);
+    let undo1 = pos_accounting::random_undo_for_test(&mut rng);
+    let undo2 = pos_accounting::random_undo_for_test(&mut rng);
+    let undo3 = pos_accounting::random_undo_for_test(&mut rng);
+    let undo3_2 = pos_accounting::random_undo_for_test(&mut rng);
 
     let block_id_1: Id<Block> = H256::random_using(&mut rng).into();
     let block_id_2: Id<Block> = H256::random_using(&mut rng).into();
@@ -42,22 +51,25 @@ fn connect_txs_in_hierarchy_default(#[case] seed: Seed) {
     let tx_id_2: Id<Transaction> = H256::random_using(&mut rng).into();
     let tx_id_3: Id<Transaction> = H256::random_using(&mut rng).into();
 
-    let expected_block_undo_1 = CachedTokensBlockUndo::new(BTreeMap::from([
-        (tx_id_1, TxUndo::new(vec![undo1.clone()])),
-        (tx_id_2, TxUndo::new(vec![undo2.clone()])),
-    ]))
+    let expected_block_undo_1 = CachedPoSBlockUndo::new(
+        None,
+        BTreeMap::from([
+            (tx_id_1, TxUndo::new(vec![undo1.clone()])),
+            (tx_id_2, TxUndo::new(vec![undo2.clone()])),
+        ]),
+    )
     .unwrap();
 
-    let expected_block_undo_2 = CachedTokensBlockUndo::new(BTreeMap::from([(
-        tx_id_3,
-        TxUndo::new(vec![undo3.clone(), undo3_2.clone()]),
-    )]))
+    let expected_block_undo_2 = CachedPoSBlockUndo::new(
+        None,
+        BTreeMap::from([(tx_id_3, TxUndo::new(vec![undo3.clone(), undo3_2.clone()]))]),
+    )
     .unwrap();
 
     let mut store = mock::MockStore::new();
     store.expect_get_best_block_for_utxos().return_const(Ok(H256::zero().into()));
     store
-        .expect_set_tokens_accounting_undo_data()
+        .expect_set_pos_accounting_undo_data()
         .with(
             eq(TransactionSource::Chain(block_id_1)),
             eq(expected_block_undo_1),
@@ -65,7 +77,7 @@ fn connect_txs_in_hierarchy_default(#[case] seed: Seed) {
         .times(1)
         .return_const(Ok(()));
     store
-        .expect_set_tokens_accounting_undo_data()
+        .expect_set_pos_accounting_undo_data()
         .with(
             eq(TransactionSource::Chain(block_id_2)),
             eq(expected_block_undo_2),
@@ -85,7 +97,7 @@ fn connect_txs_in_hierarchy_default(#[case] seed: Seed) {
     let mut base_verifier = TransactionVerifier::new(&store, &chain_config);
 
     base_verifier
-        .tokens_accounting_block_undo
+        .pos_accounting_block_undo
         .add_tx_undo(
             TransactionSource::Chain(block_id_1),
             tx_id_1,
@@ -94,7 +106,7 @@ fn connect_txs_in_hierarchy_default(#[case] seed: Seed) {
         .unwrap();
 
     base_verifier
-        .tokens_accounting_block_undo
+        .pos_accounting_block_undo
         .add_tx_undo(
             TransactionSource::Chain(block_id_1),
             tx_id_2,
@@ -103,7 +115,7 @@ fn connect_txs_in_hierarchy_default(#[case] seed: Seed) {
         .unwrap();
 
     base_verifier
-        .tokens_accounting_block_undo
+        .pos_accounting_block_undo
         .add_tx_undo(
             TransactionSource::Chain(block_id_2),
             tx_id_3,
@@ -130,10 +142,10 @@ fn connect_txs_in_hierarchy_disposable(#[case] seed: Seed) {
 
     let chain_config = ConfigBuilder::test_chain().build();
 
-    let undo1 = tokens_accounting::random_undo_for_test(&mut rng);
-    let undo2 = tokens_accounting::random_undo_for_test(&mut rng);
-    let undo3 = tokens_accounting::random_undo_for_test(&mut rng);
-    let undo3_2 = tokens_accounting::random_undo_for_test(&mut rng);
+    let undo1 = pos_accounting::random_undo_for_test(&mut rng);
+    let undo2 = pos_accounting::random_undo_for_test(&mut rng);
+    let undo3 = pos_accounting::random_undo_for_test(&mut rng);
+    let undo3_2 = pos_accounting::random_undo_for_test(&mut rng);
 
     let block_id_1: Id<Block> = H256::random_using(&mut rng).into();
     let block_id_2: Id<Block> = H256::random_using(&mut rng).into();
@@ -142,22 +154,25 @@ fn connect_txs_in_hierarchy_disposable(#[case] seed: Seed) {
     let tx_id_2: Id<Transaction> = H256::random_using(&mut rng).into();
     let tx_id_3: Id<Transaction> = H256::random_using(&mut rng).into();
 
-    let expected_block_undo_1 = CachedTokensBlockUndo::new(BTreeMap::from([
-        (tx_id_1, TxUndo::new(vec![undo1.clone()])),
-        (tx_id_2, TxUndo::new(vec![undo2.clone()])),
-    ]))
+    let expected_block_undo_1 = CachedPoSBlockUndo::new(
+        None,
+        BTreeMap::from([
+            (tx_id_1, TxUndo::new(vec![undo1.clone()])),
+            (tx_id_2, TxUndo::new(vec![undo2.clone()])),
+        ]),
+    )
     .unwrap();
 
-    let expected_block_undo_2 = CachedTokensBlockUndo::new(BTreeMap::from([(
-        tx_id_3,
-        TxUndo::new(vec![undo3.clone(), undo3_2.clone()]),
-    )]))
+    let expected_block_undo_2 = CachedPoSBlockUndo::new(
+        None,
+        BTreeMap::from([(tx_id_3, TxUndo::new(vec![undo3.clone(), undo3_2.clone()]))]),
+    )
     .unwrap();
 
     let mut store = mock::MockStore::new();
     store.expect_get_best_block_for_utxos().return_const(Ok(H256::zero().into()));
     store
-        .expect_set_tokens_accounting_undo_data()
+        .expect_set_pos_accounting_undo_data()
         .with(
             eq(TransactionSource::Chain(block_id_1)),
             eq(expected_block_undo_1),
@@ -165,7 +180,7 @@ fn connect_txs_in_hierarchy_disposable(#[case] seed: Seed) {
         .times(1)
         .return_const(Ok(()));
     store
-        .expect_set_tokens_accounting_undo_data()
+        .expect_set_pos_accounting_undo_data()
         .with(
             eq(TransactionSource::Chain(block_id_2)),
             eq(expected_block_undo_2),
@@ -186,7 +201,7 @@ fn connect_txs_in_hierarchy_disposable(#[case] seed: Seed) {
 
     let mut verifier2 = base_verifier.derive_child();
     verifier2
-        .tokens_accounting_block_undo
+        .pos_accounting_block_undo
         .add_tx_undo(
             TransactionSource::Chain(block_id_1),
             tx_id_1,
@@ -199,7 +214,7 @@ fn connect_txs_in_hierarchy_disposable(#[case] seed: Seed) {
 
     let mut verifier3 = base_verifier.derive_child();
     verifier3
-        .tokens_accounting_block_undo
+        .pos_accounting_block_undo
         .add_tx_undo(
             TransactionSource::Chain(block_id_1),
             tx_id_2,
@@ -208,7 +223,7 @@ fn connect_txs_in_hierarchy_disposable(#[case] seed: Seed) {
         .unwrap();
 
     verifier3
-        .tokens_accounting_block_undo
+        .pos_accounting_block_undo
         .add_tx_undo(
             TransactionSource::Chain(block_id_2),
             tx_id_3,
@@ -231,20 +246,22 @@ fn connect_txs_in_hierarchy_twice(#[case] seed: Seed) {
 
     let chain_config = ConfigBuilder::test_chain().build();
 
-    let undo1 = tokens_accounting::random_undo_for_test(&mut rng);
+    let undo1 = pos_accounting::random_undo_for_test(&mut rng);
 
     let block_id: Id<Block> = H256::random_using(&mut rng).into();
 
     let tx_id: Id<Transaction> = H256::random_using(&mut rng).into();
 
-    let expected_block_undo =
-        CachedTokensBlockUndo::new(BTreeMap::from([(tx_id, TxUndo::new(vec![undo1.clone()]))]))
-            .unwrap();
+    let expected_block_undo = CachedPoSBlockUndo::new(
+        None,
+        BTreeMap::from([(tx_id, TxUndo::new(vec![undo1.clone()]))]),
+    )
+    .unwrap();
 
     let mut store = mock::MockStore::new();
     store.expect_get_best_block_for_utxos().return_const(Ok(H256::zero().into()));
     store
-        .expect_set_tokens_accounting_undo_data()
+        .expect_set_pos_accounting_undo_data()
         .with(
             eq(TransactionSource::Chain(block_id)),
             eq(expected_block_undo),
@@ -264,7 +281,7 @@ fn connect_txs_in_hierarchy_twice(#[case] seed: Seed) {
     let mut base_verifier = TransactionVerifier::new(&store, &chain_config);
 
     base_verifier
-        .tokens_accounting_block_undo
+        .pos_accounting_block_undo
         .add_tx_undo(
             TransactionSource::Chain(block_id),
             tx_id,
@@ -273,15 +290,15 @@ fn connect_txs_in_hierarchy_twice(#[case] seed: Seed) {
         .unwrap();
 
     // it's an error to add the same undo via the same verifier
-    let result = base_verifier.tokens_accounting_block_undo.add_tx_undo(
+    let result = base_verifier.pos_accounting_block_undo.add_tx_undo(
         TransactionSource::Chain(block_id),
         tx_id,
         TxUndo::new(vec![undo1.clone()]),
     );
     assert_eq!(
         result.unwrap_err(),
-        ConnectTransactionError::TokensAccountingBlockUndoError(
-            tokens_accounting::BlockUndoError::UndoAlreadyExists(tx_id)
+        ConnectTransactionError::AccountingBlockUndoError(
+            pos_accounting::BlockUndoError::UndoAlreadyExists(tx_id)
         )
     );
 
@@ -290,11 +307,90 @@ fn connect_txs_in_hierarchy_twice(#[case] seed: Seed) {
     let mut derived_verifier = base_verifier.derive_child();
 
     derived_verifier
-        .tokens_accounting_block_undo
+        .pos_accounting_block_undo
         .add_tx_undo(
             TransactionSource::Chain(block_id),
             tx_id,
             TxUndo::new(vec![undo1.clone()]),
+        )
+        .unwrap();
+
+    let consumed_derived_verifier = derived_verifier.consume().unwrap();
+    flush::flush_to_storage(&mut base_verifier, consumed_derived_verifier).unwrap();
+
+    let consumed_base_verifier = base_verifier.consume().unwrap();
+    flush::flush_to_storage(&mut store, consumed_base_verifier).unwrap();
+}
+
+#[rstest]
+#[trace]
+#[case(Seed::from_entropy())]
+fn connect_reward_in_hierarchy_twice(#[case] seed: Seed) {
+    let mut rng = test_utils::random::make_seedable_rng(seed);
+
+    let chain_config = ConfigBuilder::test_chain().build();
+
+    let undo1 = pos_accounting::random_undo_for_test(&mut rng);
+
+    let block_id: Id<Block> = H256::random_using(&mut rng).into();
+
+    let expected_block_undo = CachedPoSBlockUndo::new(
+        Some(BlockRewardUndo::new(vec![undo1.clone()])),
+        BTreeMap::new(),
+    )
+    .unwrap();
+
+    let mut store = mock::MockStore::new();
+    store.expect_get_best_block_for_utxos().return_const(Ok(H256::zero().into()));
+    store
+        .expect_set_pos_accounting_undo_data()
+        .with(
+            eq(TransactionSource::Chain(block_id)),
+            eq(expected_block_undo),
+        )
+        .times(1)
+        .return_const(Ok(()));
+    store.expect_batch_write().times(1).return_const(Ok(()));
+    store
+        .expect_batch_write_tokens_data()
+        .times(1)
+        .return_const(Ok(TokensAccountingDeltaUndoData::new()));
+    store
+        .expect_batch_write_delta()
+        .times(1)
+        .return_const(Ok(DeltaMergeUndo::new()));
+
+    let mut base_verifier = TransactionVerifier::new(&store, &chain_config);
+
+    base_verifier
+        .pos_accounting_block_undo
+        .add_reward_undo(
+            TransactionSource::Chain(block_id),
+            BlockRewardUndo::new(vec![undo1.clone()]),
+        )
+        .unwrap();
+
+    // it's an error to add the same undo via the same verifier
+    let result = base_verifier.pos_accounting_block_undo.add_reward_undo(
+        TransactionSource::Chain(block_id),
+        BlockRewardUndo::new(vec![undo1.clone()]),
+    );
+    assert_eq!(
+        result.unwrap_err(),
+        ConnectTransactionError::AccountingBlockUndoError(
+            pos_accounting::BlockUndoError::UndoAlreadyExistsForReward
+        )
+    );
+
+    // It's an not an error to add the same undo via derived verifier.
+    // There is no way to distinguish this case with current approach because info is not fetched on add.
+    let mut derived_verifier = base_verifier.derive_child();
+
+    derived_verifier
+        .pos_accounting_block_undo
+        .add_reward_undo(
+            TransactionSource::Chain(block_id),
+            BlockRewardUndo::new(vec![undo1]),
         )
         .unwrap();
 
@@ -313,10 +409,10 @@ fn disconnect_txs_in_hierarchy_default(#[case] seed: Seed) {
 
     let chain_config = ConfigBuilder::test_chain().build();
 
-    let undo1 = tokens_accounting::random_undo_for_test(&mut rng);
-    let undo2 = tokens_accounting::random_undo_for_test(&mut rng);
-    let undo3 = tokens_accounting::random_undo_for_test(&mut rng);
-    let undo3_2 = tokens_accounting::random_undo_for_test(&mut rng);
+    let undo1 = pos_accounting::random_undo_for_test(&mut rng);
+    let undo2 = pos_accounting::random_undo_for_test(&mut rng);
+    let undo3 = pos_accounting::random_undo_for_test(&mut rng);
+    let undo3_2 = pos_accounting::random_undo_for_test(&mut rng);
 
     let block_id_1: Id<Block> = H256::random_using(&mut rng).into();
     let block_id_2: Id<Block> = H256::random_using(&mut rng).into();
@@ -328,36 +424,39 @@ fn disconnect_txs_in_hierarchy_default(#[case] seed: Seed) {
     let block_undo_db = BTreeMap::from_iter([
         (
             TransactionSource::Chain(block_id_1),
-            CachedTokensBlockUndo::new(BTreeMap::from([
-                (tx_id_1, TxUndo::new(vec![undo1.clone()])),
-                (tx_id_2, TxUndo::new(vec![undo2.clone()])),
-            ]))
+            CachedPoSBlockUndo::new(
+                None,
+                BTreeMap::from([
+                    (tx_id_1, TxUndo::new(vec![undo1.clone()])),
+                    (tx_id_2, TxUndo::new(vec![undo2.clone()])),
+                ]),
+            )
             .unwrap(),
         ),
         (
             TransactionSource::Chain(block_id_2),
-            CachedTokensBlockUndo::new(BTreeMap::from([(
-                tx_id_3,
-                TxUndo::new(vec![undo3.clone(), undo3_2.clone()]),
-            )]))
+            CachedPoSBlockUndo::new(
+                None,
+                BTreeMap::from([(tx_id_3, TxUndo::new(vec![undo3.clone(), undo3_2.clone()]))]),
+            )
             .unwrap(),
         ),
     ]);
 
     let fetch_block_undo =
-        |tx_source| -> Result<Option<CachedTokensBlockUndo>, ConnectTransactionError> {
+        |tx_source| -> Result<Option<CachedPoSBlockUndo>, ConnectTransactionError> {
             Ok(block_undo_db.get(&tx_source).cloned())
         };
 
     let mut store = mock::MockStore::new();
     store.expect_get_best_block_for_utxos().return_const(Ok(H256::zero().into()));
     store
-        .expect_del_tokens_accounting_undo_data()
+        .expect_del_pos_accounting_undo_data()
         .with(eq(TransactionSource::Chain(block_id_1)))
         .times(1)
         .return_const(Ok(()));
     store
-        .expect_del_tokens_accounting_undo_data()
+        .expect_del_pos_accounting_undo_data()
         .with(eq(TransactionSource::Chain(block_id_2)))
         .times(1)
         .return_const(Ok(()));
@@ -375,7 +474,7 @@ fn disconnect_txs_in_hierarchy_default(#[case] seed: Seed) {
 
     // Undo tx2
     base_verifier
-        .tokens_accounting_block_undo
+        .pos_accounting_block_undo
         .take_tx_undo(
             &TransactionSource::Chain(block_id_1),
             &tx_id_2,
@@ -385,7 +484,7 @@ fn disconnect_txs_in_hierarchy_default(#[case] seed: Seed) {
 
     // Undo tx1
     base_verifier
-        .tokens_accounting_block_undo
+        .pos_accounting_block_undo
         .take_tx_undo(
             &TransactionSource::Chain(block_id_1),
             &tx_id_1,
@@ -394,7 +493,7 @@ fn disconnect_txs_in_hierarchy_default(#[case] seed: Seed) {
         .unwrap();
 
     base_verifier
-        .tokens_accounting_block_undo
+        .pos_accounting_block_undo
         .take_tx_undo(
             &TransactionSource::Chain(block_id_2),
             &tx_id_3,
@@ -414,10 +513,10 @@ fn disconnect_txs_in_hierarchy_disposable(#[case] seed: Seed) {
 
     let chain_config = ConfigBuilder::test_chain().build();
 
-    let undo1 = tokens_accounting::random_undo_for_test(&mut rng);
-    let undo2 = tokens_accounting::random_undo_for_test(&mut rng);
-    let undo3 = tokens_accounting::random_undo_for_test(&mut rng);
-    let undo3_2 = tokens_accounting::random_undo_for_test(&mut rng);
+    let undo1 = pos_accounting::random_undo_for_test(&mut rng);
+    let undo2 = pos_accounting::random_undo_for_test(&mut rng);
+    let undo3 = pos_accounting::random_undo_for_test(&mut rng);
+    let undo3_2 = pos_accounting::random_undo_for_test(&mut rng);
 
     let block_id_1: Id<Block> = H256::random_using(&mut rng).into();
     let block_id_2: Id<Block> = H256::random_using(&mut rng).into();
@@ -426,36 +525,39 @@ fn disconnect_txs_in_hierarchy_disposable(#[case] seed: Seed) {
     let tx_id_2: Id<Transaction> = H256::random_using(&mut rng).into();
     let tx_id_3: Id<Transaction> = H256::random_using(&mut rng).into();
 
-    let block_undo_1 = CachedTokensBlockUndo::new(BTreeMap::from([
-        (tx_id_1, TxUndo::new(vec![undo1.clone()])),
-        (tx_id_2, TxUndo::new(vec![undo2.clone()])),
-    ]))
+    let block_undo_1 = CachedPoSBlockUndo::new(
+        None,
+        BTreeMap::from([
+            (tx_id_1, TxUndo::new(vec![undo1.clone()])),
+            (tx_id_2, TxUndo::new(vec![undo2.clone()])),
+        ]),
+    )
     .unwrap();
 
-    let block_undo_2 = CachedTokensBlockUndo::new(BTreeMap::from([(
-        tx_id_3,
-        TxUndo::new(vec![undo3.clone(), undo3_2.clone()]),
-    )]))
+    let block_undo_2 = CachedPoSBlockUndo::new(
+        None,
+        BTreeMap::from([(tx_id_3, TxUndo::new(vec![undo3.clone(), undo3_2.clone()]))]),
+    )
     .unwrap();
 
     let mut store = mock::MockStore::new();
     store.expect_get_best_block_for_utxos().return_const(Ok(H256::zero().into()));
     store
-        .expect_get_tokens_accounting_undo()
+        .expect_get_pos_accounting_undo()
         .with(eq(TransactionSource::Chain(block_id_1)))
         .return_const(Ok(Some(block_undo_1.clone())));
     store
-        .expect_get_tokens_accounting_undo()
+        .expect_get_pos_accounting_undo()
         .with(eq(TransactionSource::Chain(block_id_2)))
         .return_const(Ok(Some(block_undo_2.clone())));
 
     store
-        .expect_del_tokens_accounting_undo_data()
+        .expect_del_pos_accounting_undo_data()
         .with(eq(TransactionSource::Chain(block_id_1)))
         .times(1)
         .return_const(Ok(()));
     store
-        .expect_del_tokens_accounting_undo_data()
+        .expect_del_pos_accounting_undo_data()
         .with(eq(TransactionSource::Chain(block_id_2)))
         .times(1)
         .return_const(Ok(()));
@@ -475,13 +577,13 @@ fn disconnect_txs_in_hierarchy_disposable(#[case] seed: Seed) {
 
     {
         let fetch_block_undo =
-            |tx_source| -> Result<Option<CachedTokensBlockUndo>, ConnectTransactionError> {
-                Ok(base_verifier.get_tokens_accounting_undo(tx_source)?)
+            |tx_source| -> Result<Option<CachedPoSBlockUndo>, ConnectTransactionError> {
+                Ok(base_verifier.get_pos_accounting_undo(tx_source)?)
             };
 
         // Undo tx2
         derived_verifier
-            .tokens_accounting_block_undo
+            .pos_accounting_block_undo
             .take_tx_undo(
                 &TransactionSource::Chain(block_id_1),
                 &tx_id_2,
@@ -498,12 +600,12 @@ fn disconnect_txs_in_hierarchy_disposable(#[case] seed: Seed) {
 
     {
         let fetch_block_undo =
-            |tx_source| -> Result<Option<CachedTokensBlockUndo>, ConnectTransactionError> {
-                Ok(base_verifier.get_tokens_accounting_undo(tx_source)?)
+            |tx_source| -> Result<Option<CachedPoSBlockUndo>, ConnectTransactionError> {
+                Ok(base_verifier.get_pos_accounting_undo(tx_source)?)
             };
 
         derived_verifier
-            .tokens_accounting_block_undo
+            .pos_accounting_block_undo
             .take_tx_undo(
                 &TransactionSource::Chain(block_id_1),
                 &tx_id_1,
@@ -519,12 +621,12 @@ fn disconnect_txs_in_hierarchy_disposable(#[case] seed: Seed) {
     let mut derived_verifier = base_verifier.derive_child();
     {
         let fetch_block_undo =
-            |tx_source| -> Result<Option<CachedTokensBlockUndo>, ConnectTransactionError> {
-                Ok(base_verifier.get_tokens_accounting_undo(tx_source)?)
+            |tx_source| -> Result<Option<CachedPoSBlockUndo>, ConnectTransactionError> {
+                Ok(base_verifier.get_pos_accounting_undo(tx_source)?)
             };
 
         derived_verifier
-            .tokens_accounting_block_undo
+            .pos_accounting_block_undo
             .take_tx_undo(
                 &TransactionSource::Chain(block_id_2),
                 &tx_id_3,
@@ -548,8 +650,8 @@ fn connect_disconnect_for_mempool(#[case] seed: Seed) {
 
     let chain_config = ConfigBuilder::test_chain().build();
 
-    let undo1 = tokens_accounting::random_undo_for_test(&mut rng);
-    let undo2 = tokens_accounting::random_undo_for_test(&mut rng);
+    let undo1 = pos_accounting::random_undo_for_test(&mut rng);
+    let undo2 = pos_accounting::random_undo_for_test(&mut rng);
 
     let tx_id_1: Id<Transaction> = H256::random_using(&mut rng).into();
     let tx_id_2: Id<Transaction> = H256::random_using(&mut rng).into();
@@ -560,7 +662,7 @@ fn connect_disconnect_for_mempool(#[case] seed: Seed) {
     let mut base_verifier = TransactionVerifier::new(&store, &chain_config);
 
     base_verifier
-        .tokens_accounting_block_undo
+        .pos_accounting_block_undo
         .add_tx_undo(
             TransactionSource::Mempool,
             tx_id_1,
@@ -569,7 +671,7 @@ fn connect_disconnect_for_mempool(#[case] seed: Seed) {
         .unwrap();
 
     base_verifier
-        .tokens_accounting_block_undo
+        .pos_accounting_block_undo
         .add_tx_undo(
             TransactionSource::Mempool,
             tx_id_2,
@@ -581,12 +683,12 @@ fn connect_disconnect_for_mempool(#[case] seed: Seed) {
 
     {
         let fetch_block_undo =
-            |tx_source| -> Result<Option<CachedTokensBlockUndo>, ConnectTransactionError> {
-                Ok(base_verifier.get_tokens_accounting_undo(tx_source)?)
+            |tx_source| -> Result<Option<CachedPoSBlockUndo>, ConnectTransactionError> {
+                Ok(base_verifier.get_pos_accounting_undo(tx_source)?)
             };
 
         derived_verifier
-            .tokens_accounting_block_undo
+            .pos_accounting_block_undo
             .take_tx_undo(&TransactionSource::Mempool, &tx_id_2, fetch_block_undo)
             .unwrap();
     }
@@ -595,10 +697,10 @@ fn connect_disconnect_for_mempool(#[case] seed: Seed) {
     flush::flush_to_storage(&mut base_verifier, consumed_derived_verifier).unwrap();
 
     let fetch_block_undo =
-        |_| -> Result<Option<CachedTokensBlockUndo>, ConnectTransactionError> { Ok(None) };
+        |_| -> Result<Option<CachedPoSBlockUndo>, ConnectTransactionError> { Ok(None) };
 
     base_verifier
-        .tokens_accounting_block_undo
+        .pos_accounting_block_undo
         .take_tx_undo(&TransactionSource::Mempool, &tx_id_1, fetch_block_undo)
         .unwrap();
 }
@@ -611,7 +713,7 @@ fn connect_disconnect_connect_for_mempool(#[case] seed: Seed) {
 
     let chain_config = ConfigBuilder::test_chain().build();
 
-    let undo1 = tokens_accounting::random_undo_for_test(&mut rng);
+    let undo1 = pos_accounting::random_undo_for_test(&mut rng);
 
     let tx_id_1: Id<Transaction> = H256::random_using(&mut rng).into();
 
@@ -622,7 +724,7 @@ fn connect_disconnect_connect_for_mempool(#[case] seed: Seed) {
 
     // Connect a transaction in base
     base_verifier
-        .tokens_accounting_block_undo
+        .pos_accounting_block_undo
         .add_tx_undo(
             TransactionSource::Mempool,
             tx_id_1,
@@ -635,12 +737,12 @@ fn connect_disconnect_connect_for_mempool(#[case] seed: Seed) {
     // Disconnect a transaction in derived and flush
     {
         let fetch_block_undo =
-            |tx_source| -> Result<Option<CachedTokensBlockUndo>, ConnectTransactionError> {
-                Ok(base_verifier.get_tokens_accounting_undo(tx_source)?)
+            |tx_source| -> Result<Option<CachedPoSBlockUndo>, ConnectTransactionError> {
+                Ok(base_verifier.get_pos_accounting_undo(tx_source)?)
             };
 
         derived_verifier
-            .tokens_accounting_block_undo
+            .pos_accounting_block_undo
             .take_tx_undo(&TransactionSource::Mempool, &tx_id_1, fetch_block_undo)
             .unwrap();
     }
@@ -652,7 +754,7 @@ fn connect_disconnect_connect_for_mempool(#[case] seed: Seed) {
     let mut derived_verifier = base_verifier.derive_child();
 
     derived_verifier
-        .tokens_accounting_block_undo
+        .pos_accounting_block_undo
         .add_tx_undo(
             TransactionSource::Mempool,
             tx_id_1,
