+++ chainstate/tx-verifier/src/transaction_verifier/tokens_accounting_undo_cache/mod.rs
@@ -152,3 +152,6 @@ impl TokensAccountingBlockUndoCache {
         Ok(())
     }
 }
+
+#[cfg(test)]
+mod tests;
