+++ chainstate/tx-verifier/src/transaction_verifier/utxos_undo_cache/mod.rs
@@ -247,3 +247,6 @@ impl UtxosBlockUndoCache {
         Ok(())
     }
 }
+
+#[cfg(test)]
+mod tests;
